+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+<!--
+Copyright (C) 2020-2022 Arm Limited or its affiliates and Contributors. All rights reserved.
+SPDX-License-Identifier: Apache-2.0
+-->
 # Changelog
 
 <!--
@@ -12,6 +16,193 @@ beta releases are not included in this history. For a full list of all releases,
 This project was forked from version 1.7.4 of [mbed-tools-ci-scripts](https://github.com/ARMmbed/mbed-tools-ci-scripts)
 
 [//]: # (begin_release_notes)
+
+"2.8.5" (2022-03-22)
+====================
+
+Bugfixes
+--------
+
+- [Golang] Disabling code documentation generation until [issue](https://github.com/go101/golds/issues/26) with `golds` is fixed (#202203221358)
+
+
+"2.8.4" (2022-03-22)
+====================
+
+Bugfixes
+--------
+
+- [Golang] Fixing how golds is called (#202203221309)
+
+
+"2.8.3" (2022-03-22)
+====================
+
+Bugfixes
+--------
+
+- Dependency upgrade: towncrier-21.9.0 (#202202241608)
+- [GOLANG] Fix golds version so that it works on go > 1.18 required by latest goreleaser (#202203220937)
+- Change `towncrier` CLI command as no longer backward compatible (#202203221036)
+
+
+"2.8.2" (2022-02-21)"2.8.2" (2022-02-21)
+====================
+
+Bugfixes
+--------
+
+- Fixing markupsafe version to solve Jinja's(dependency issue)[https://github.com/pallets/markupsafe/issues/284] (#202202211807)
+
+
+"2.8.1" (2022-02-17)"2.8.1" (2022-02-17)
+====================
+
+Bugfixes
+--------
+
+- [GO] Fixed [golds version](https://github.com/go101/golds/issues/26) by fixing the version to v0.4.1 (#202202171446)
+
+
+Improved Documentation
+----------------------
+
+- Documented how to update the secret checkers (#202202171830)
+
+
+Misc
+----
+
+- #202202171829
+
+
+"2.8.0" (2021-08-26)"2.8.0" (2021-08-26)
+====================
+
+Features
+--------
+
+- Made release tagging language configurable. (#202108251834)
+
+
+"2.7.2" (2021-08-25)"2.7.2" (2021-08-25)
+====================
+
+Bugfixes
+--------
+
+- [GO] Fixed goreleaser version tag issue (#202108250248)
+
+
+"2.7.1" (2021-08-24)"2.7.1" (2021-08-24)
+====================
+
+Bugfixes
+--------
+
+- Fixed issue with goreleaser requiring tag to be checked out (#202108242248)
+
+
+"2.7.0" (2021-08-24)"2.7.0" (2021-08-24)
+====================
+
+Features
+--------
+
+- Added a stash repository step (#202108242058)
+- Added a language specific option to clean repository before packaging (#202108242145)
+
+
+"2.6.2" (2021-08-24)"2.6.2" (2021-08-24)
+====================
+
+Bugfixes
+--------
+
+- Fixed repository cleanup (#202108241912)
+
+
+"2.6.1" (2021-08-24)"2.6.1" (2021-08-24)
+====================
+
+Bugfixes
+--------
+
+- Fixed Golang packaging (#202108241747)
+
+
+"2.6.0" (2021-08-24)"2.6.0" (2021-08-24)
+====================
+
+Features
+--------
+
+- Added a clean repository step (#202108241231)
+
+
+"2.5.1" (2021-08-24)"2.5.1" (2021-08-24)
+====================
+
+Bugfixes
+--------
+
+- Fixed tag and release (#202108240918)
+
+
+"2.5.0" (2021-08-13)"2.5.0" (2021-08-13)
+====================
+
+Features
+--------
+
+- Added a [pre-commit](https://pre-commit.com) hook for ease of use in other repositories (#202108131306)
+
+
+Bugfixes
+--------
+
+- Dependency upgrade: towncrier-21.3.0 (#202108110250)
+
+
+"2.4.0" (2021-08-12)
+====================
+
+Features
+--------
+
+- Added a golang plugin (#202108112203)
+
+
+Improved Documentation
+----------------------
+
+- Improved configuration documentation (#202108111714)
+
+
+"2.3.1" (2021-08-11)
+====================
+
+Bugfixes
+--------
+
+- Made `cd-determine-version` idempotent (#202108111624)
+
+
+"2.3.0" (2021-08-11)
+====================
+
+Features
+--------
+
+- Added a command to determine project's version (#202108111437)
+
+
+Bugfixes
+--------
+
+- Dependency upgrade: licenseheaders-lt-0.8.9 (#202108110242)
+- Fix bug in determining the new release/beta version (#202108111347)
+
 
 "2.2.4" (2021-08-11)
 ====================